% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{a_upset_plot}
\alias{a_upset_plot}
\title{a_upset_plot}
\usage{
a_upset_plot(
  list,
  nintersects = 40,
  order.intersect.by = c("size", "name"),
  order.set.by = c("size", "name")
)
}
\arguments{
\item{list}{a list of sets}

\item{nintersects}{number of intersects}

\item{order.intersect.by}{one of 'size' or 'name'}

\item{order.set.by}{one of 'size' or 'name'}
}
\value{
an upset plot
}
\description{
Plot a upset plot
}
\details{
This function generate a upset plot by creating a composite plot which contains subplots generated by ggplot2.
}
\examples{
 list = list(A = sample(LETTERS, 20),
             B = sample(LETTERS, 22),
             C = sample(LETTERS, 14),
             D = sample(LETTERS, 30, replace = TRUE))
 a_upset_plot(list)
 a_upset_plot(list, order.intersect.by = "name")
 a_upset_plot(list, nintersects = 6)
}
