% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplotExtra-package.R
\docType{package}
\name{aplotExtra-package}
\alias{aplotExtra}
\alias{aplotExtra-package}
\title{aplotExtra: Creating Composite Plots using 'aplot'}
\description{
Many complex plots are actually composite plots, such as 'oncoplot', 'funkyheatmap', 'upsetplot', etc. We can produce subplots using 'ggplot2' and combine them to create composite plots using 'aplot'. In this way, it is easy to customize these complex plots, by adding, deleting or modifying subplots in the final plot. This package provides a set of utilities to help users to create subplots and complex plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/aplotExtra}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID}) [contributor]
  \item Chun-Hui Gao \email{gaospecial@gmail.com} (\href{https://orcid.org/0000-0002-1445-7939}{ORCID}) [contributor]
  \item Shensuo Li \email{lishensuo@163.com} (\href{https://orcid.org/0009-0006-3229-3700}{ORCID}) [contributor]
}

}
\keyword{internal}
