% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.r
\name{oncoplot_components}
\alias{oncoplot_components}
\title{Get reusable oncoplot components}
\usage{
oncoplot_components(maf, genes = 20)
}
\arguments{
\item{maf}{A MAF object.}

\item{genes}{The gene names or number of top mutated genes.}
}
\value{
A list with components:
\describe{
\item{data}{Tidy oncoplot matrix.}
\item{main}{Main mutation tile plot.}
\item{sample}{Sample-level top annotation plot.}
\item{gene}{Gene-level side annotation plot.}
\item{colors}{Named mutation-class color vector.}
}
}
\description{
Build the reusable panels and tidy data used by \code{\link[=oncoplot]{oncoplot()}}.
This function is intended for downstream packages that want to assemble
an oncoplot with additional tracks while reusing the aplotExtra mutation
matrix, sample burden, gene burden, and mutation color mapping.
}
