% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upsetplot2}
\alias{upsetplot2}
\title{upsetplot2}
\usage{
upsetplot2(
  list,
  nintersects = NULL,
  order.intersect.by = c("size", "name"),
  order.set.by = c("size", "name")
)
}
\arguments{
\item{list}{a list of sets}

\item{nintersects}{number of intersects. If NULL, all intersections will show.}

\item{order.intersect.by}{one of 'size' or 'name'}

\item{order.set.by}{one of 'size' or 'name'}
}
\value{
an upset plot
}
\description{
Plot a upset plot
}
\details{
This function generate a upset plot by creating a composite plot which contains subplots generated by ggplot2.
}
\examples{
 list = list(A = sample(LETTERS, 20),
             B = sample(LETTERS, 22),
             C = sample(LETTERS, 14),
             D = sample(LETTERS, 30, replace = TRUE))
 upsetplot2(list)
 upsetplot2(list, order.intersect.by = "name")
 upsetplot2(list, nintersects = 6)
}
